refactor cloudtasks API
