+++ apis/cloudtasks/v1alpha1/groupversion_info.go
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 // +kubebuilder:object:generate=true
-// +groupName=tasks.cnrm.cloud.google.com
+// +groupName=cloudtasks.cnrm.cloud.google.com
 package v1alpha1
 
 import (
@@ -23,7 +23,7 @@ import (
 
 var (
 	// GroupVersion is group version used to register these objects
-	GroupVersion = schema.GroupVersion{Group: "tasks.cnrm.cloud.google.com", Version: "v1alpha1"}
+	GroupVersion = schema.GroupVersion{Group: "cloudtasks.cnrm.cloud.google.com", Version: "v1alpha1"}
 
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme
 	SchemeBuilder = &scheme.Builder{GroupVersion: GroupVersion}
