+++ config/installbundle/components/clusterroles/cnrm_admin.yaml
@@ -1280,18 +1280,6 @@ rules:
   - patch
   - delete
 - apiGroups:
-  - tasks.cnrm.cloud.google.com
-  resources:
-  - '*'
-  verbs:
-  - get
-  - list
-  - watch
-  - create
-  - update
-  - patch
-  - delete
-- apiGroups:
   - tpu.cnrm.cloud.google.com
   resources:
   - '*'
