+++ config/installbundle/components/clusterroles/cnrm_viewer.yaml
@@ -855,14 +855,6 @@ rules:
   - list
   - watch
 - apiGroups:
-  - tasks.cnrm.cloud.google.com
-  resources:
-  - '*'
-  verbs:
-  - get
-  - list
-  - watch
-- apiGroups:
   - tpu.cnrm.cloud.google.com
   resources:
   - '*'
