+++ mockgcp/mockcloudtasks/service.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package mocktasks
+package mockcloudtasks
 
 // +tool:mockgcp-service
 // http.host: workflows.googleapis.com
