+++ pkg/test/resourcefixture/testdata/basic/cloudtasks/v1alpha1/cloudtasksqueue/queueautogen/_generated_object_queueautogen.golden.yaml
@@ -1,4 +1,4 @@
-apiVersion: tasks.cnrm.cloud.google.com/v1alpha1
+apiVersion: cloudtasks.cnrm.cloud.google.com/v1alpha1
 kind: TasksQueue
 metadata:
   finalizers:
