+++ pkg/test/resourcefixture/testdata/basic/cloudtasks/v1alpha1/cloudtasksqueue/queueautogen/create.yaml
@@ -1,0 +1,20 @@
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+apiVersion: cloudtasks.cnrm.cloud.google.com/v1alpha1
+kind: TasksQueue
+metadata:
+  name: cloudtasksqueue-${uniqueId}
+spec:
+  location: us-central1
+  projectRef:
+    external: ${projectId}
+  resourceID: cloudtasksqueue${uniqueId}
