+++ pkg/test/resourcefixture/testdata/basic/cloudtasks/v1alpha1/cloudtasksqueue/queueautogen/update.yaml
@@ -1,6 +1,3 @@
-# Copyright 2023 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
@@ -12,7 +9,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-apiVersion: tasks.cnrm.cloud.google.com/v1alpha1
+apiVersion: cloudtasks.cnrm.cloud.google.com/v1alpha1
 kind: TasksQueue
 metadata:
   name: cloudtasksqueue-${uniqueId}
