--- pkg/test/resourcefixture/testdata/basic/tasks/v1alpha1/tasksqueue/tasksqueue-minimal/update.yaml
@@ -1,9 +1,0 @@
-apiVersion: tasks.cnrm.cloud.google.com/v1alpha1
-kind: TasksQueue
-metadata:
-  name: tasksqueue-minimal-${uniqueId}
-spec:
-  projectRef:
-    external: ${projectId}
-  locationID: us-west2
-  description: "Updated description"
